RF: re-found tests for cmdline and moved under cmdline module. merged with just added
